Create website.py
This is just an extra step to run the hand tracking across a website build on your own. To use this file, the module flask must be installed in your python interpreter. Do note that an index.html in necessary to ensure the website run in the first place and a localhost should be initialized beforehand (xampp)